Python/requirements.txt
numpy==1.18.0
beautifulsoup4==4.12.2
cython
shapely==1.7.1
proj==0.2.0
pyshp==2.1.3
setuptools==40.8.0
six==1.12.0
cartopy==0.18.0
feedparser==6.0.10
geopandas==0.10.2
geoplot==0.5.1
geopy==2.3.0
matplotlib==3.3.3
nltk==3.8.1
objgraph==3.6.0
pandas==1.1.4
Pillow==9.5.0
pyephem==9.99
pytz==2023.3
rclone==0.4.4
Requests==2.31.0
scipy==1.5.4
skyfield==1.35
opencv-python
opencv-python-headless
tk